Display changes, new commands to zopflipng and more
Display block splitter counters only on verbosity level 5 as it is the same for iterations progression in compatiblity mode (--t0),
Improve block splitter estimate costs display,
Change ratio to display compression ratio not percentage of original size,
Display progress indicator for FINISHED blocks instead of started - makes it feel more real on what the current compression progress is,
Dynamically convert from B to KB and MB for Data Left,
Expose new block splitter parameters to ZopfliPNG
Fix ZopfliPNG blinking percentage progress in genetic algorithm

+++ src/zopfli/squeeze.c
@@ -586,7 +586,7 @@ zfloat ZopfliLZ77Optimal(ZopfliBlockState *s,
       iterations->iteration = i;
       iterations->cost = (int)cost;
     } else if (s->options->verbose>4 || (s->options->verbose>2 && cost < bestcost)) {
-      fprintf(stderr, "Iteration %d: %d bit      \r", i, (int) cost);
+      fprintf(stderr, "Iteration %d: %lu bit      \r", i, (unsigned long)cost);
     }
     if (cost < bestcost) {
       if(s->options->numthreads) {
